requirements.txt
Flask==2.3.0
Flask-Mail==0.9.1
Flask-Login==0.6.2
Werkzeug==2.3.0
reportlab==4.0.4
Pillow==10.0.0
itsdangerous==2.1.2
python-dotenv==1.0.0
gunicorn==21.2.0